text: Centered textbox height calculation
A simple hotfix, which changes the calculation of the bounding box of a
text shape. With this most text is now centered according to the cap
height, which is the perceived text center.

note: needed to multiply font metric measurement by 10 in order to get
subpixel resolution. This fixes a few of by one errors during
measurement. Sadly it seems like that increasing the factor doesn't
result in more precise values. i.e. measuring a font multiplied by 100
results in very wrong results. So *10 seems to be the sweat spot.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>